+++ src/index.tsx
@@ -2,6 +2,7 @@ import "./index.css";
 import { createRoot } from "react-dom/client";
 import { Main } from "./main";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import ThemeProvider from "./contexts/ThemeProvider";
 
 const container = document.querySelector("#root");
 const root = createRoot(container!);
@@ -9,6 +10,8 @@ const queryClient = new QueryClient();
 
 root.render(
   <QueryClientProvider client={queryClient}>
-    <Main />
+    <ThemeProvider>
+      <Main />
+    </ThemeProvider>
   </QueryClientProvider>
 );
